Add support for specifying PATH for subprocesses.
Kas would very much like to use "/usr/sbin:/usr/bin:/sbin:/bin" as the
PATH when it runs Bitbake. Furthermore, it doesn't set a PATH at all
when it runs Git, causing Python to only look for the git binary in
/usr/bin and /bin. This ought to be fine, but some systems store things
differently, causing Kas to be unable to find Git or to execute the
wrong Git.

I have kludged around this to account for hosts that aren't set up the
way Siemens envisions by allowing the PATH to be set by reading
KAS_HOST_PATH from the environment. If KAS_HOST_PATH is not set,
"/usr/sbin:/usr/bin:/sbin:/bin" will be used both for Bitbake and Git
runs.

Since the main reason for spurning the caller's path is to make builds
more hermetic, I set KAS_HOST_PATH to the caller's path with any entries
in the user's home directory stripped out:

  KAS_HOST_PATH=$(echo "$PATH" | tr : "\n" | grep -v "^$HOME" | tr "\n" :)

You may choose to be more or less strict, depending on your environment.